refactor(core): fix viewer  module imports
