+++ projects/dsp-ui/src/lib/search/advanced-search/advanced-search.component.ts
@@ -23,7 +23,7 @@ import {
 import { Subscription } from 'rxjs';
 import { NotificationService } from '../../action/services/notification.service';
 import { DspApiConnectionToken } from '../../core/core.module';
-import { SearchParams } from '../../viewer';
+import { SearchParams } from '../../viewer/';
 import { GravsearchGenerationService } from '../services/gravsearch-generation.service';
 import { Properties, SelectPropertyComponent } from './select-property/select-property.component';
 import { PropertyWithValue } from './select-property/specify-property-value/operator';
