+++ projects/dsp-ui/src/lib/search/advanced-search/select-property/specify-property-value/search-date-value/search-date-value.component.spec.ts
@@ -1,16 +1,15 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SearchDateValueComponent } from './search-date-value.component';
-import { Component, Inject, OnInit, ViewChild } from '@angular/core';
-import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatInputModule } from '@angular/material/input';
-import { CalendarDate, CalendarPeriod, GregorianCalendarDate, JDNPeriod } from 'jdnconvertiblecalendar';
-import { ValueLiteral } from '../operator';
+import { Component, Inject, OnInit, ViewChild } from '@angular/core';
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
 import { MatDatepickerModule } from '@angular/material/datepicker';
-import { DspViewerModule } from '../../../../../viewer';
+import { MatInputModule } from '@angular/material/input';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { CalendarDate, CalendarPeriod, GregorianCalendarDate } from 'jdnconvertiblecalendar';
+import { DspViewerModule } from '../../../../../viewer/viewer.module';
+import { ValueLiteral } from '../operator';
+import { SearchDateValueComponent } from './search-date-value.component';
 
 /**
  * Test host component to simulate parent component.
