+++ projects/dsp-ui/src/lib/search/advanced-search/select-property/specify-property-value/search-date-value/search-date-value.component.ts
@@ -1,9 +1,9 @@
 import { Component, Inject, Input, OnDestroy, OnInit } from '@angular/core';
-import { PropertyValue, Value, ValueLiteral } from '../operator';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { Constants } from '@dasch-swiss/dsp-js';
 import { JDNConvertibleCalendar } from 'jdnconvertiblecalendar';
-import { CalendarHeaderComponent } from '../../../../../viewer';
+import { CalendarHeaderComponent } from '../../../../../viewer/values/date-value/calendar-header/calendar-header.component';
+import { PropertyValue, Value, ValueLiteral } from '../operator';
 
 // https://stackoverflow.com/questions/45661010/dynamic-nested-reactive-form-expressionchangedafterithasbeencheckederror
 const resolvedPromise = Promise.resolve(null);
